tools/audio-pipeline/requirements.txt
google-genai>=1.0.0
boto3>=1.35.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydub>=0.25.1
